+++ src/components/properties/EffectsSection.vue
@@ -5,7 +5,8 @@ import NumberField from '@/components/inputs/NumberField.vue'
 import IconButton from '@/components/ui/IconButton.vue'
 import PanelSection from '@/components/ui/panel/PanelSection.vue'
 import Tip from '@/components/ui/Tip.vue'
-import { PropertyListRoot, vTestId, useEffectsControls, useI18n } from '@open-pencil/vue'
+import PropertyListRoot from '@/components/properties/PropertyListRoot.vue'
+import { vTestId, useEffectsControls, useI18n } from '@open-pencil/vue'
 
 import { colorToCSS } from '@open-pencil/core/color'